feat(AdminAuthorizationValidator): API별 권한 검증 추가

+++ src/main/java/com/classroom/manager/user/presentation/AdminController.java
@@ -4,34 +4,40 @@ import com.classroom.manager.user.application.AdminLoginService;
 import com.classroom.manager.user.application.AdminRegisterService;
 import com.classroom.manager.user.application.dto.AdminLoginRequest;
 import com.classroom.manager.user.application.dto.RegisterAdminRequest;
+import com.classroom.manager.user.infra.security.AdminAuthorizationValidator;
+import com.classroom.manager.user.infra.security.annotation.Auth;
+import com.classroom.manager.user.infra.security.dto.TokenPayLoad;
 import com.classroom.manager.user.presentation.dto.LoginResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
-@Controller
+@RestController
 @RequestMapping("/admins")
 @RequiredArgsConstructor
 public class AdminController {
 
     private final AdminRegisterService adminRegisterService;
     private final AdminLoginService adminLoginService;
+    private final AdminAuthorizationValidator adminAuthorizationValidator;
 
     @PostMapping
-    public ResponseEntity<Void> registerAdmin(@RequestBody RegisterAdminRequest registerAdminRequest) {
+    public ResponseEntity<Void> registerAdmin(@Auth TokenPayLoad tokenPayLoad, @RequestBody RegisterAdminRequest registerAdminRequest) {
+        adminAuthorizationValidator.checkCanRegister(tokenPayLoad, registerAdminRequest.authorization());
         adminRegisterService.register(registerAdminRequest);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping
-    public ResponseEntity<Void> deleteAdmin(@RequestBody RegisterAdminRequest registerAdminRequest) {
+    public ResponseEntity<Void> deleteAdmin(@Auth TokenPayLoad tokenPayLoad, @RequestBody RegisterAdminRequest registerAdminRequest) {
+        adminAuthorizationValidator.checkIsAdminOrHigher(tokenPayLoad);
         adminRegisterService.delete(registerAdminRequest);
         return ResponseEntity.ok().build();
     }
 
     @PatchMapping
-    public ResponseEntity<Void> updateAdmin(@RequestBody RegisterAdminRequest registerAdminRequest) {
+    public ResponseEntity<Void> updateAdmin(@Auth TokenPayLoad tokenPayLoad, @RequestBody RegisterAdminRequest registerAdminRequest) {
+        adminAuthorizationValidator.checkIsSuperAdmin(tokenPayLoad);
         adminRegisterService.suspend(registerAdminRequest);
         return ResponseEntity.ok().build();
     }
